Switch from GET to POST
Since geojson can get really large (100K+ characters), they will not fit
in the limit of a query string. Thus we move all parameters to be POSTed
instead.

In order for direct JSON to be captured by Scala, we define case classes
for each end-point, and use the entity directive instead of parameters.