Return array of values instead of map
Since the values in the geoprocessing service are in Meters, but the
front-end uses Feet, to return a map of the keys with odd looking
decimal values in Meters would be confusing and unnecessary. Instead, it
is simpler for us to simply send back an array of percentages,which
correspond to whatever flood levels were specified during input.

Refs azavea/usace-flood-model#71