Visualize Depth
We move to a rendering which shows how much a cell will be flooded at a
given flood level, instead of showing at what flood level will the cell
flood. This will be the difference between the flood level, and the
relative elevation of the cell (which is its elevation minus the minimum
elevation).

Also, we add color breaks specified by the client. We repeat the final
value to a very high maximum (virtually infinite flooding) to ensure
that the last color persists even in values beyond that of the last
break 15'. Without it, GeoTrellis defaults to the value of the first
break, which is very light, and misrepresentative of the true amount of
flooding in the cell.

Refs #15