Return 200 for valid empty tiles
In cases when a tile is valid, but has no values, previously we were
returning a 404. This would flood (no pun intended) the browser with
numerous error messages, which were inaccurate since the tiles were
correctly empty.

Now we respond with a 200 and an empty body, which should work just as
well without the browser showing so many errors.

Refs #14