Add Docker container and upgrade GeoTrellis
Builds the geoprocessing module in a Docker container so that it can be
easily integrated into usace-levee-safety-api project, which uses Docker.

This implementation is largely based off of
https://github.com/azavea/usace-program-analysis-geoprocessing.

GeoTrellis is also upgraded here because the version that this module
previously relied on is no longer available.

Note that, at this point, the container does not successfully build
because of Scala compiler errors related to version 1.0.0 of GeoTrellis.
These errors will be addressed in a follow-up issue.

Refs https://github.com/azavea/usace-levee-safety-api/issues/3